CatalogConstans class moved from out of the exceptions package
