+++ src/main/java/org/fogbowcloud/saps/engine/core/catalog/CatalogConstants.java
@@ -1,4 +1,4 @@
-package org.fogbowcloud.saps.engine.core.catalog.exceptions;
+package org.fogbowcloud.saps.engine.core.catalog;
 
 public class CatalogConstants {
     public static final int UNLIMITED = -1;
