+++ src/main/java/org/fogbowcloud/saps/engine/utils/retry/catalog/GetProcessingTasksRetry.java
@@ -3,7 +3,7 @@ package org.fogbowcloud.saps.engine.utils.retry.catalog;
 import java.util.List;
 
 import org.fogbowcloud.saps.engine.core.catalog.Catalog;
-import org.fogbowcloud.saps.engine.core.catalog.exceptions.CatalogConstants;
+import org.fogbowcloud.saps.engine.core.catalog.CatalogConstants;
 import org.fogbowcloud.saps.engine.core.model.SapsImage;
 import org.fogbowcloud.saps.engine.core.model.enums.ImageTaskState;
